Added Linear Search

+++ algonation/src/main/java/com/algonation/search/searchers/ISearcher.java
@@ -2,5 +2,5 @@ package com.algonation.search.searchers;
 
 public interface ISearcher {
 
-	public Double search(Double[] array, Double element);
+	public int search(Double[] array, Double element);
 }
